// frontend/src/app/interfaces/exercise.interface.ts
interface Exercice {
  titulo?: string;
  descripcion: string;
  imagen?: string;
  parteCuerpo?: string;
  musculo?: string;
  consejos?: string[];
  ejercicioAlternativo?: string;
  video?: string;
}

export interface Exercise {
  titulo?: string;
  descripcion: string;
  imagen?: string;
  parteCuerpo?: string;
  musculo?: string;
  consejos?: string[];
  ejercicioAlternativo?: string;
  video?: string;
}